Update: updated main structure on header
