+++ src/App.js
@@ -1,9 +1,14 @@
 import './App.css';
+import Header from './components/Header';
+import TopHeader from './components/TopHeader';
+import Hero from './components/Hero';
 
 function App() {
   return (
     <div className="App">
-    
+      <TopHeader/>
+      <Header/>
+      <Hero/>
     </div>
   );
 }
